Feature/api testing

+++ PortfolioService/src/main/java/com/SwitchBoard/PortfolioService/Service/Portfolio/Impl/FileServiceImpl.java
@@ -1,6 +1,5 @@
 package com.SwitchBoard.PortfolioService.Service.Portfolio.Impl;
 
-
 import com.SwitchBoard.PortfolioService.Service.Portfolio.FileService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,7 +8,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.PutObjectAclRequest;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 
 import java.io.IOException;
@@ -17,8 +15,8 @@ import java.util.UUID;
 
 @Service
 public class FileServiceImpl implements FileService {
-    private static final Logger log = LoggerFactory.getLogger(FileServiceImpl.class);
 
+    private static final Logger log = LoggerFactory.getLogger(FileServiceImpl.class);
     private final S3Client s3Client;
 
     @Value("${aws.s3.bucket}")
@@ -33,39 +31,35 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public String uploadImage(String path, MultipartFile file) throws IOException {
-        log.info("FileServiceImpl :: uploadImage :: uploading image: {}", file.getOriginalFilename());
+        log.info("Uploading file: {} to S3 path: {}", file.getOriginalFilename(), path);
 
-        // Random file name for uniqueness
         String fileName = UUID.randomUUID() + "_" + file.getOriginalFilename();
         String key = path + "/" + fileName;
 
-        // Create put request with public-read ACL
-        PutObjectRequest putObjectRequest = PutObjectRequest.builder()
+        PutObjectRequest request = PutObjectRequest.builder()
                 .bucket(bucket)
                 .key(key)
                 .contentType(file.getContentType())
                 .acl("public-read")
                 .build();
 
-        log.info("FileServiceImpl :: uploadImage :: uploading to S3 with key: {}", key);
-        s3Client.putObject(putObjectRequest, RequestBody.fromBytes(file.getBytes()));
+        s3Client.putObject(request, RequestBody.fromBytes(file.getBytes()));
 
-        // Generate permanent public URL
         String publicUrl = "https://" + bucket + ".s3." + region + ".amazonaws.com/" + key;
+        log.info("File uploaded successfully. Public URL: {}", publicUrl);
 
-        log.info("FileServiceImpl :: uploadImage :: completed. Public URL: {}", publicUrl);
         return publicUrl;
     }
 
-
     @Override
     public void deleteImage(String fileUrl) {
-        log.info("FileServiceImpl :: deleteImage :: deleting image: {}", fileUrl);
+        if (fileUrl == null || !fileUrl.contains(".com/")) {
+            log.warn("Invalid file URL, skipping delete: {}", fileUrl);
+            return;
+        }
 
-        // Extract key from URL
         String key = fileUrl.substring(fileUrl.indexOf(".com/") + 5);
-
+        log.info("Deleting file from S3 with key: {}", key);
         s3Client.deleteObject(builder -> builder.bucket(bucket).key(key));
-        log.info("FileServiceImpl :: deleteImage :: deleted image from S3: {}", key);
     }
 }
